refactor(office): extract conversation panel with chat-view style messages
- Extract ConversationPanel, OverlayBubbles, ArchivePanel into components/office/
- Replace plain text messages with Markdown rendering and avatar initials
- Use PromptInput with ArrowUp/Square button matching chat-view.tsx layout
- Display "You" for user messages, use sidebar-consistent agent colors
- Remove Conversations title and SSE status footer
- Add office-themed inline code styling for markdown
